[Feat/#79] 할 일/다가오는 일정 고도화 (#80)
* [Feat/#79] 스케줄 바텀시트 스낵바 호스트 추가

* [Feat/#79] 스케줄 삭제 로딩 상태 추가

* [Fix/#79] 스케줄 바텀시트 필드 포커스 미이동 이슈 해결
- focusManager로는 제어가 안 되어 focusRequest 사용

* [Fix/#79] 스케줄 바텀시트 bottomPadding 이동
- Column에 bottomPadding이 있으니 어색해보여 LargeButton으로 이동

* [Fix/#79] 텍스트필드 힌트텍스트 입력 중 미노출되게 수정

* [Feat/#79] todo/schedule 수정 시 변경사항 없으면 버튼 비활성화 & 삭제 로딩 중 호출 방지

* [Refactor/#79] todo/schedule 로딩 중 깜박이지 않게 수정

* [Feat/#79] schedule 헬퍼 아이콘 버튼 추가

* [Fix/#79] todo/schedule 리스트 키 추가
- 순서 변경 시 팝업이 다른 아이템에 붙을 위험이 있어 id값을 key로 추가

+++ app/src/main/java/com/swyp/firsttodo/presentation/todo/TodoViewModel.kt
@@ -160,8 +160,6 @@ class TodoViewModel
                 }
 
         suspend fun getTodos() {
-            updateState { copy(todos = Async.Loading(this.todos.getDataOrNull())) }
-
             todoRepository.getTodos()
                 .onSuccess { data ->
                     val categories = uiState.value.categories
@@ -200,8 +198,6 @@ class TodoViewModel
         }
 
         fun getSchedules() {
-            updateState { copy(schedules = Async.Loading(this.schedules.getDataOrNull())) }
-
             viewModelScope.launch {
                 scheduleRepository.getSchedules()
                     .onSuccess { list ->
@@ -291,6 +287,9 @@ class TodoViewModel
                         title = todoUiModel.title,
                         category = todoUiModel.category,
                         labelColor = todoUiModel.labelColor,
+                        originalTitle = todoUiModel.title,
+                        originalCategory = todoUiModel.category,
+                        originalLabelColor = todoUiModel.labelColor,
                     ),
                 )
             }
@@ -328,7 +327,12 @@ class TodoViewModel
                     scheduleBottomSheetState = Async.Init,
                     editingSchedule = editingSchedule.copy(
                         scheduleId = scheduleUiModel.scheduleId,
+                        title = scheduleUiModel.title,
+                        date = scheduleUiModel.rawDate,
                         category = scheduleUiModel.category,
+                        originalTitle = scheduleUiModel.title,
+                        originalDate = scheduleUiModel.rawDate,
+                        originalCategory = scheduleUiModel.category,
                     ),
                 )
             }
@@ -379,6 +383,7 @@ class TodoViewModel
         }
 
         fun onDeleteConfirm() {
+            if (uiState.value.deleteState is Async.Loading) return
             when (uiState.value.delRequestedType) {
                 DeleteDialogType.Todo -> deleteTodo()
                 DeleteDialogType.Schedule -> deleteSchedule()
@@ -416,7 +421,10 @@ class TodoViewModel
                                 throwable.snackbarMsg()
                             }
 
-                            else -> return@onFailure
+                            else -> {
+                                updateState { copy(deleteState = Async.Init) }
+                                return@onFailure
+                            }
                         }
 
                         sendEffect(TodoSideEffect.ShowSnackbar(message))
@@ -427,24 +435,37 @@ class TodoViewModel
         private fun deleteSchedule() {
             val scheduleId = uiState.value.delRequestedId ?: return
 
+            updateState { copy(deleteState = Async.Loading()) }
+
             viewModelScope.launch {
                 scheduleRepository.deleteSchedule(scheduleId)
                     .onSuccess {
-                        getSchedules()
-                        updateState { copy(delRequestedId = null) }
+                        updateState { copy(delRequestedId = null, deleteState = Async.Success(Unit)) }
                         sendEffect(TodoSideEffect.ShowSnackbar("다가오는 일정이 삭제되었습니다."))
+                        getSchedules()
                     }
                     .onFailure { throwable ->
                         val message = when (throwable) {
                             is ScheduleError.ScheduleNotFound -> {
-                                updateState { copy(delRequestedId = null) }
+                                updateState {
+                                    copy(
+                                        delRequestedId = null,
+                                        deleteState = Async.Success(Unit),
+                                    )
+                                }
                                 getSchedules()
                                 "이미 삭제된 일정입니다."
                             }
 
-                            is ApiError -> throwable.snackbarMsg()
+                            is ApiError -> {
+                                updateState { copy(deleteState = Async.Init) }
+                                throwable.snackbarMsg()
+                            }
 
-                            else -> return@launch
+                            else -> {
+                                updateState { copy(deleteState = Async.Init) }
+                                return@launch
+                            }
                         }
                         sendEffect(TodoSideEffect.ShowSnackbar(message))
                     }
@@ -653,7 +674,10 @@ class TodoViewModel
                                 throwable.snackbarMsg()
                             }
 
-                            else -> return@launch
+                            else -> {
+                                updateState { copy(scheduleBottomSheetState = Async.Init) }
+                                return@launch
+                            }
                         }
                         sendEffect(TodoSideEffect.ShowSnackbar(message))
                     }
